services/ad-orchestration/meta-ads-fetcher/requirements.txt
facebook-business==19.0.3
schedule==1.2.2
python-dotenv==1.0.1
requests==2.32.3
supabase==2.15.1
google-genai>=1.0.0
google-cloud-vision>=3.7.0
